Detached board logic from project logic
***
frontend
